--- fw/io.go
@@ -1,6 +1,0 @@
-package fw
-
-import "io"
-
-type StdIn io.Reader
-type StdOut io.Writer
