+++ fw/io/io.go
@@ -1,0 +1,6 @@
+package io
+
+import "io"
+
+type Input io.Reader
+type Output io.Writer
